Move pins around to put outputs on PORTA
The high-impedance input mode of the PORTB pins has insufficiently high
impedance to use as a "low" state for buttons 2 and 3.  Shift the pins
around so that all outputs are on PORTA, where the input mode resistance
is higher.

Driving the pins high (for up/down/left/right/fire) or low (for buttons
two and three) can cause problems for the 8-bit Commodores.  Switching
the signal tristates between input and output instead avoids this.