Simplify button state logic
Make a new flag called TWO_BUTTON in the controller state to track if a
second button is present, simplifying the code that disconnects the
buttons in Atari mode and allowing it to use the same tristate logic as
the Commodores.

Remove the SIX_BUTTON flags since they're not used.