Add six-button support with autofire
The select line toggle sequence to read the X, Y, Z, and MODE buttons is
timing sensitive, and is too slow for the 8bitdo receiver with the PIC
system clock at 500kHz (the default) or 1MHz. At 2MHz, it works.

Without multiplexing on the computer side there's no way to use the
extra buttons directly, but autofire is useful.  Holding down X, Y, or Z
respectively pulses the A, B, or C button at around 12Hz.