Add experimental SNACK support
Hold down the MODE button to switch to SNACK mode.

SNACK is a project that brings SNES controllers to Atari, and it has a
"quite clever ;) but also easy to use" mapping for all SNES buttons to
the Atari.

The button layout is adapted to the SEGA controller thusly:

- A and B are swapped
- X and Y are swapped
- Z and C are the left and right shoulder buttons, respectively
- MODE is SELECT

The 8BitDo M30 controller has shoulder buttons that map to Z and C.

Link to the SNACK discussion thread on the AtariAge forums:
https://forums.atariage.com/topic/323171-snack-snes-atari-controler-kit/